Log: Update.
Sorry for skipping updates to the log for the last 3 months! 🙏

I was trying to finish a surprisingly MASSIVE upgrade to the log,
and was hoping to avoid what I thought would be a complex rebase,
which turned out to be very simple after all. 🤦

Right now I am upgrading my JavaScript skills, in part to help me
finish said upgrade, which I will finish and publish in Q1 2025.

Apologies to Anonymous 3 in particular!! 🙏
Your heroism is appreciated!! 💪